Move YAJL code into a subdirectory.
