+++ json.cc
@@ -43,9 +43,9 @@
 #include "streams.h"
 #include "unparse.h"
 #include "utils.h"
-#include "yajl_gen.h"
-#include "yajl_lex.h"
-#include "yajl_parse.h"
+#include "yajl/yajl_gen.h"
+#include "yajl/yajl_lex.h"
+#include "yajl/yajl_parse.h"
 
 /*
   Handle many modes of mapping between JSON and internal MOO types.
