+++ yajl/yajl_alloc.c
