+++ yajl/yajl_buf.c
