+++ yajl/yajl_bytestack.h
