+++ yajl/yajl_encode.c
