+++ yajl/yajl_gen.c
