+++ yajl/yajl_lex.c
