+++ yajl/yajl_version.c
